+++ src/main/java/com/heavenscode/rac/repository/BillingserviceoptionvaluesRepository.java
@@ -10,4 +10,4 @@ import org.springframework.stereotype.Repository;
 @SuppressWarnings("unused")
 @Repository
 public interface BillingserviceoptionvaluesRepository
-    extends JpaRepository<Billingserviceoptionvalues, Long>, JpaSpecificationExecutor<Billingserviceoptionvalues> {}
+    extends JpaRepository<Billingserviceoptionvalues, Integer>, JpaSpecificationExecutor<Billingserviceoptionvalues> {}
