+++ src/main/java/com/heavenscode/rac/repository/WorkshopVehicleWorkListRepository.java
@@ -9,4 +9,4 @@ import org.springframework.stereotype.Repository;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface WorkshopVehicleWorkListRepository extends JpaRepository<WorkshopVehicleWorkList, Long> {}
+public interface WorkshopVehicleWorkListRepository extends JpaRepository<WorkshopVehicleWorkList, Integer> {}
